weather_source/requirements.txt
apscheduler
fastapi
uvicorn
redis
celery
requests
python-dotenv
pymongo
confluent-kafka